Replaced old collect_all_text with new one
